requirements.txt
# Tamil Handwritten OCR System - Requirements
# Dependencies for converting Tamil handwritten images to digital text with cultural context

# Core OCR and Image Processing
opencv-python>=4.8.0        # Image preprocessing and enhancement
easyocr>=1.7.0              # Primary OCR engine for Tamil text
pytesseract>=0.3.10         # Secondary OCR engine
Pillow>=10.0.0              # Image manipulation and format handling
numpy>=1.24.0               # Numerical operations for image arrays

# Enhancement integration
google-generativeai>=0.3.0  # Gemini integration for enhanced cultural analysis
requests>=2.28.0            # HTTP requests for API calls

# Image Enhancement and Computer Vision  
scikit-image>=0.20.0        # Advanced image processing algorithms
scipy>=1.10.0               # Mathematical operations for image enhancement
matplotlib>=3.7.0           # Image visualization and debugging

# Database and Text Processing
sqlite3                     # Built-in Python SQLite support (no pip install needed)
difflib                     # Built-in text similarity (no pip install needed)
fuzzywuzzy>=0.18.0         # Fuzzy string matching for verse comparison
python-Levenshtein>=0.20.0  # Fast string distance calculations

# Web Interface and API
flask>=2.3.0               # Web framework for user interface
flask-cors>=4.0.0          # Cross-origin resource sharing
werkzeug>=2.3.0            # WSGI utilities
jinja2>=3.1.0              # Template engine

# File Handling and Utilities
pathlib                    # Built-in path handling (no pip install needed)
base64                     # Built-in encoding (no pip install needed)
io                         # Built-in I/O operations (no pip install needed)
json                       # Built-in JSON handling (no pip install needed)
logging                    # Built-in logging (no pip install needed)
time                       # Built-in time operations (no pip install needed)
os                         # Built-in OS interface (no pip install needed)
re                         # Built-in regular expressions (no pip install needed)
typing                     # Built-in type hints (no pip install needed)

# Development and Testing
pytest>=7.0.0             # Testing framework
pytest-cov>=4.0.0         # Test coverage reporting
black>=23.0.0              # Code formatting
flake8>=6.0.0              # Code linting

# Tamil Language Support
indic-transliteration>=2.3.0  # Tamil script handling
tamil>=1.0.0               # Tamil language utilities (if available)

# Performance and Monitoring
psutil>=5.9.0              # System resource monitoring
tqdm>=4.65.0               # Progress bars for processing

# Optional: Advanced ML/model integrations (for future enhancements)
# tensorflow>=2.13.0       # Deep learning for advanced OCR
# torch>=2.0.0             # PyTorch for neural networks
# transformers>=4.30.0     # Hugging Face transformers for NLP

# Note: Install Tesseract OCR separately on your system:
# Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki
# Linux: sudo apt-get install tesseract-ocr tesseract-ocr-tam
# macOS: brew install tesseract

# Installation command:
# pip install -r requirements.txt